refactor: use the same reflection logic
