+++ velocity/src/main/java/host/bloom/ab/velocity/VelocityPlugin.java
@@ -6,8 +6,6 @@ import com.velocitypowered.api.event.proxy.ProxyInitializeEvent;
 import com.velocitypowered.api.plugin.Plugin;
 import com.velocitypowered.api.plugin.PluginContainer;
 import com.velocitypowered.api.proxy.ProxyServer;
-import com.velocitypowered.proxy.VelocityServer;
-import com.velocitypowered.proxy.network.ConnectionManager;
 import dev.geri.konfig.util.InvalidConfigurationException;
 import host.bloom.ab.common.AbstractPlugin;
 import host.bloom.ab.common.config.Config;
@@ -17,7 +15,6 @@ import host.bloom.ab.common.utils.Scheduler;
 import host.bloom.ab.common.utils.UpdateChecker;
 
 import java.io.IOException;
-import java.lang.reflect.Field;
 import java.nio.file.Paths;
 
 @Plugin(id = "@id@", name = "@name@", version = "@version@", description = "@description@")
@@ -37,7 +34,6 @@ public class VelocityPlugin implements AbstractPlugin {
 
     @Subscribe
     public void onProxyInitialization(ProxyInitializeEvent event) {
-
         // Initialize the plugin
         this.plugin = server.getPluginManager().ensurePluginContainer(this);
 
@@ -61,21 +57,12 @@ public class VelocityPlugin implements AbstractPlugin {
         // We can either use the built-in event or our own channel listener
         if (this.config.catchRawConnections) {
 
-            // Impossible to get exception, only for new version of Velocity.
+            // Initialize the login hook channel
             try {
-                Class<?> velocityServerClass = VelocityServer.class;
-                Field cmField = velocityServerClass.getDeclaredField("cm");
-                cmField.setAccessible(true);
-
-                // Get the connection manager
-                ConnectionManager connectionManager = (ConnectionManager) cmField.get(server);
-
-                // Initialize the login hook channel
-                VelocityLoginHookChannel loginHookChannel = new VelocityLoginHookChannel(this, connectionManager.getServerChannelInitializer().get());
-                connectionManager.getServerChannelInitializer().set(loginHookChannel);
+                new VelocityLoginHookChannel(this, server);
                 return;
-
-            } catch (IllegalAccessException | NoSuchFieldException exception) {
+            } catch (Exception exception) {
+                // Impossible to get exception, only for new version of Velocity.
                 this.getABLogger().error("Unable to initialize raw connection catcher, using built-in events: " + exception.getMessage());
                 exception.printStackTrace();
             }
@@ -108,4 +95,5 @@ public class VelocityPlugin implements AbstractPlugin {
     public Config getABConfig() {
         return this.config;
     }
+
 }
